refactor: use StructuredData server component with XSS sanitization
