+++ frontend/src/app/layout.tsx
@@ -2,9 +2,11 @@ import "./globals.css";
 
 import type { Metadata } from "next";
 import Link from "next/link";
+import Script from "next/script";
 
 import { Footer } from "@/components/Footer";
 import HeaderMenu from "@/components/HeaderMenu";
+import { StructuredData } from "@/components/StructuredData";
 import { ToastProvider } from "@/components/Toast";
 import { getHeaderColor } from "@/utils/environment";
 
@@ -43,45 +45,8 @@ const getBaseUrl = (): URL => {
 // metadataBase を取得
 const metadataBase = getBaseUrl();
 
-export async function generateMetadata(): Promise<Metadata & { jsonLd?: any }> {
+export async function generateMetadata(): Promise<Metadata> {
   const baseUrl = getBaseUrl();
-  const siteUrlString = baseUrl.toString();
-
-  const websiteSchema = {
-    "@context": "https://schema.org",
-    "@type": "WebSite",
-    name: "Smart Food Logger",
-    url: siteUrlString,
-  };
-
-  const softwareAppSchema = {
-    "@context": "https://schema.org",
-    "@type": "SoftwareApplication",
-    name: "Smart Food Logger",
-    applicationCategory: "HealthApplication",
-    operatingSystem: "Browser",
-    offers: {
-      "@type": "Offer",
-      price: "0",
-      priceCurrency: "JPY",
-    },
-    description: SITE_DESCRIPTION,
-    author: {
-      "@type": "Organization",
-      name: "Smart Food Logger Team",
-      url: siteUrlString,
-    },
-    screenshot: `${siteUrlString}images/ogp-image.png`.replace(
-      /([^:]\/)\/+/g,
-      "$1",
-    ),
-    featureList: [
-      "食事検索や面倒な栄養計算・入力が一切不要",
-      "写真やチャットでAIに伝えるだけの圧倒的に楽な記録体験",
-      "生成されたJSONをコピペするだけのシンプル操作",
-      "Fitbitへの自動連携（完全無料）",
-    ],
-  };
 
   return {
     metadataBase: baseUrl,
@@ -99,7 +64,6 @@ export async function generateMetadata(): Promise<Metadata & { jsonLd?: any }> {
       description: SITE_DESCRIPTION,
       images: [OG_IMAGE],
     },
-    jsonLd: [websiteSchema, softwareAppSchema],
   };
 }
 
@@ -110,10 +74,12 @@ export default function RootLayout({
 }) {
   const environment = process.env.APP_ENVIRONMENT;
   const headerBgClass = getHeaderColor(environment);
+  const baseUrl = getBaseUrl().toString();
 
   return (
     <html lang="ja">
       <body>
+        <StructuredData baseUrl={baseUrl} />
         {/* Preconnect hints for critical third-party domains to improve initial load performance */}
         {/* Next.js will hoist these to the <head> */}
         <link rel="preconnect" href="https://apis.google.com" />
